refactor(filters): split chain.rs into SPL submodules
Decompose the 1024-line `chain.rs` into a `chain/` module directory with
concern-named submodules to keep each file focused and within LoC limits:

- `config.rs`  - `DirMergeConfig` and modifier application
- `scope.rs`   - `DirScope` (private), `DirFilterGuard`, `has_matching_rule`
- `error.rs`   - `FilterChainError` with `Display` / `Error` impls
- `mod.rs`     - `FilterChain` orchestration with push/pop semantics
- `tests.rs`   - existing tests, co-located via `#[cfg(test)] mod tests;`

Pure refactor. Public API is unchanged: re-exports in `lib.rs` remain
`pub use chain::{DirFilterGuard, DirMergeConfig, FilterChain, FilterChainError};`.
